removed the offending obfuscated tests per Ripley's suggestion:
On CRAN now.

However, one of the obfuscated tests is failing on some platforms, at least OS X and Solaris:

Running the tests in ‘tests/testthat.R’ failed.
Last 13 lines of output:
  Checking list2df : .....................................
  Checking lookup : .......
  Checking mtabulate : ....
  Checking pad : .......1........
  Checking text2color : ....
  Checking v_outer : ...............

  1. Failure(@test-pad.R#19): pad produces vectors of the correct nchar ----------
  "      May" %in% pad(month.name) isn't true

and on OS X

> pad(month.name)
 [1] "000000May" "00000July" "00000June" "0000April" "0000March" "000August"
 [7] "00January" "00October" "0December" "0February" "0November" "September"

You are relying on the behaviour of the system sprintf("%09s").  The behaviour of that modifier for %s is not mandated by either the C99 standard nor POSIX.  And ?sprintf in R says

     The following is abstracted from Kernighan and Ritchie (see
     References): however the actual implementation will follow the C99
     standard and fine details (especially the behaviour under user
     error) may depend on the platform.

     ‘0’ For numbers, pad to the field width with leading zeros.

Please submit a corrected update once all the results are in (see the CRAN policies).